archive/projects/crypto-arb-bot/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-arb-bot"
version = "0.1.0"
description = "Safe-by-default cross-exchange spot arbitrage bot with backtesting"
requires-python = ">=3.11"
dependencies = [
  "ccxt>=4.4.0",
  "numpy>=1.26",
  "pandas>=2.2",
  "pyyaml>=6.0",
]

[tool.setuptools]
packages = ["arb_bot"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
